# openapi.yaml
openapi: 3.0.0
info:
  title: API Car
  description: API para o gerenciamento de carros por atributos
  version: 1.0.0
  contact:
    name: desenvolvedores
    email: jhonatan.lima_bols@compasso.com.br
externalDocs:
  description: Documentação (README)
  url: 'https://github.com/jhonatanfariasdelima/sprint3Compasso/blob/main/README.md'
servers:
  - url: 'http://localhost:8080/api'
    description: API cars
paths:
  /cars:
    get:
      summary: Apresenta todos os veículos cadastrados na base de dados
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  chassi:
                    type: string
                  modelo:
                    type: string
                  marca:
                    type: string
                  cor:
                    type: string
                  ano_fabricacao:
                    type: string
                  valor:
                    type: number
                  dados_pagina:
                    type: object
        '401':
          description: item não encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  campo:
                    type: string
                  mensagem:
                    type: string
    post:
      summary: adiciona um novo veículo a base de dados
      requestBody:
        content:
          application/json:
            schema: 
              type: object
              properties: 
                chassi:
                  type: string
                modelo:
                  type: string
                marca:
                  type: string
                cor:
                  type: string
                valor:
                  type: number
                ano_fabricacao:
                  type: string
      responses:
        '201':
          description: cadastrado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  chassi:
                    type: string
                  modelo:
                    type: string
                  marca:
                    type: string
                  cor:
                    type: string
                  anoFabricacao:
                    type: string
                  valor:
                    type: number
        '409':
          description: duplicado (erro)
          content:
            application/json:
              schema:
                type: object
                properties:
                  campo:
                    type: string
                  erro:
                    type: string
        '400':
          description: estrutural (erro)
          content:
            application/json:
              schema:
                type: object
                properties:
                  campo:
                    type: string
                  erro:
                    type: string
                    
  /cars/:
    parameters:
      - name: chassi
        in: query
        schema:
          type: string
        requered: false
      - name: modelo
        in: query
        schema:
          type: string
        requered: false
      - name: marca
        in: query
        schema:
          type: string
        requered: false
      - name: cor
        in: query
        schema:
          type: string
        requered: false
      - name: sort
        in: query
        schema:
          type: string
        requered: false
      - name: size
        in: query
        schema:
          type: string
        requered: false
    get:
      summary: >-
        Apresenta os dados cadastrados no banco buscando por um filtro
        especifico
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  chassi:
                    type: string
                  modelo:
                    type: string
                  marca:
                    type: string
                  cor:
                    type: string
                  ano_fabricacao:
                    type: string
                  valor:
                    type: number
                    format: float  
                  dados_pagina:
                    type: object
        '401':
          description: item não encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  campo:
                    type: string
                  mensagem:
                    type: string